chore: corrections after review.
Ticket: MEN-6381
